Rename C source files according to the test suite they belong to
Implement "regression" tests in regression_*.c, the "gp" tests in
gp_*.c, and the "benchmark" tests in benchmark_*.c.
