+++ .gitignore
@@ -6,11 +6,11 @@ GTAGS
 out/*
 ta/GP_*
 host/xtest/global_platform/
-host/xtest/xtest_7000_gp.c
-host/xtest/xtest_7500.c
-host/xtest/xtest_8000.c
-host/xtest/xtest_8500.c
-host/xtest/xtest_9000.c
+host/xtest/gp_7000_gp.c
+host/xtest/gp_7500.c
+host/xtest/gp_8000.c
+host/xtest/gp_8500.c
+host/xtest/gp_9000.c
 host/xtest/adbg_case_declare.h
 host/xtest/adbg_entry_declare.h
 package/testsuite/global_platform/api_1.0/TEE*
