+++ Makefile
@@ -131,17 +131,17 @@ patch-generate-host: patch-package
 	$(q) mkdir -p ${GP_XTEST_IN_DIR} ${GP_XTEST_IN_DIR}
 	$(q)find ${CFG_GP_PACKAGE_PATH}/packages -type f -name "*.xml" -exec cp -p {} ${GP_XTEST_IN_DIR} \;
 	$(q)find ${CFG_GP_XSL_PACKAGE_PATH}/packages -type f -name "*.xsl" -exec cp -p {} ${GP_XTEST_IN_DIR} \;
-	$(call patch-xalan,TEE.xml,TEE.xsl,xtest_7000_gp.c)
-	$(call patch-xalan,TEE_DataStorage_API.xml,TEE_DataStorage_API.xsl,xtest_7500.c)
-	$(call patch-xalan,TEE_Internal_API.xml,TEE_Internal_API.xsl,xtest_8000.c)
-	$(call patch-xalan,TEE_TimeArithm_API.xml,TEE_TimeArithm_API.xsl,xtest_8500.c)
-	$(call patch-xalan,TEE_Crypto_API.xml,TEE_Crypto_API.xsl,xtest_9000.c)
+	$(call patch-xalan,TEE.xml,TEE.xsl,gp_7000.c)
+	$(call patch-xalan,TEE_DataStorage_API.xml,TEE_DataStorage_API.xsl,gp_7500.c)
+	$(call patch-xalan,TEE_Internal_API.xml,TEE_Internal_API.xsl,gp_8000.c)
+	$(call patch-xalan,TEE_TimeArithm_API.xml,TEE_TimeArithm_API.xsl,gp_8500.c)
+	$(call patch-xalan,TEE_Crypto_API.xml,TEE_Crypto_API.xsl,gp_9000.c)
 	@echo "INFO: Patch host tests"
 	# $(q)sed -i '752 c\    xtest_tee_deinit();\n' ${GP_XTEST_OUT_DIR}/xtest_7000.c
 	# $(q)sed -i '1076 c\    xtest_tee_deinit();\n' ${GP_XTEST_OUT_DIR}/xtest_8000.c
 	# $(q)sed -i '2549 c\    xtest_tee_deinit();\n' ${GP_XTEST_OUT_DIR}/xtest_8500.c
 	# $(q)sed -i '246 c\    xtest_tee_deinit();\n' ${GP_XTEST_OUT_DIR}/xtest_9000.c
-	$(call patch-file,host/xtest/xtest_9000.c,${CFG_GP_XSL_PACKAGE_PATH}/host/xtest/xtest_9000.c.patch)
+	$(call patch-file,host/xtest/gp_9000.c,${CFG_GP_XSL_PACKAGE_PATH}/host/xtest/gp_9000.c.patch)
 
 .PHONY: patch-generate-ta
 patch-generate-ta: patch-package
@@ -205,54 +205,54 @@ endef
 .PHONY: patch-filter
 patch-filter:
 	@echo "INFO: Filter some tests"
-	$(call patch-filter-one,7038,xtest_7000_gp.c)
-	$(call patch-filter-one,7522,xtest_7500.c)
-	$(call patch-filter-one,7538,xtest_7500.c)
-	$(call patch-filter-one,7540,xtest_7500.c)
-	$(call patch-filter-one,7546,xtest_7500.c)
-	$(call patch-filter-one,7557,xtest_7500.c)
-	$(call patch-filter-one,7559,xtest_7500.c)
-	$(call patch-filter-one,7577,xtest_7500.c)
-	$(call patch-filter-one,7641,xtest_7500.c)
-	$(call patch-filter-one,7642,xtest_7500.c)
-	$(call patch-filter-one,7643,xtest_7500.c)
-	$(call patch-filter-one,7644,xtest_7500.c)
-	$(call patch-filter-one,7686,xtest_7500.c)
-	$(call patch-filter-one,8025,xtest_8000.c)
-	$(call patch-filter-one,8030,xtest_8000.c)
-	$(call patch-filter-one,8058,xtest_8000.c)
-	$(call patch-filter-one,8059,xtest_8000.c)
-	$(call patch-filter-one,8066,xtest_8000.c)
-	$(call patch-filter-one,8614,xtest_8500.c)
-	$(call patch-filter-one,8643,xtest_8500.c)
-	$(call patch-filter-one,8644,xtest_8500.c)
-	$(call patch-filter-one,8673,xtest_8500.c)
-	$(call patch-filter-one,8674,xtest_8500.c)
-	$(call patch-filter-one,9001,xtest_9000.c)
-	$(call patch-filter-one,9072,xtest_9000.c)
-	$(call patch-filter-one,9073,xtest_9000.c)
-	$(call patch-filter-one,9075,xtest_9000.c)
-	$(call patch-filter-one,9079,xtest_9000.c)
-	$(call patch-filter-one,9080,xtest_9000.c)
-	$(call patch-filter-one,9082,xtest_9000.c)
-	$(call patch-filter-one,9085,xtest_9000.c)
-	$(call patch-filter-one,9086,xtest_9000.c)
-	$(call patch-filter-one,9088,xtest_9000.c)
-	$(call patch-filter-one,9090,xtest_9000.c)
-	$(call patch-filter-one,9091,xtest_9000.c)
-	$(call patch-filter-one,9093,xtest_9000.c)
-	$(call patch-filter-one,9095,xtest_9000.c)
-	$(call patch-filter-one,9096,xtest_9000.c)
-	$(call patch-filter-one,9098,xtest_9000.c)
-	$(call patch-filter-one,9099,xtest_9000.c)
-	$(call patch-filter-one,9109,xtest_9000.c)
-	$(call patch-filter-one,9110,xtest_9000.c)
-	$(call patch-filter-one,9160,xtest_9000.c)
-	$(call patch-filter-one,9174,xtest_9000.c)
-	$(call patch-filter-one,9195,xtest_9000.c)
-	$(call patch-filter-one,9196,xtest_9000.c)
-	$(call patch-filter-one,9204,xtest_9000.c)
-	$(call patch-filter-one,9239,xtest_9000.c)
+	$(call patch-filter-one,7038,gp_7000.c)
+	$(call patch-filter-one,7522,gp_7500.c)
+	$(call patch-filter-one,7538,gp_7500.c)
+	$(call patch-filter-one,7540,gp_7500.c)
+	$(call patch-filter-one,7546,gp_7500.c)
+	$(call patch-filter-one,7557,gp_7500.c)
+	$(call patch-filter-one,7559,gp_7500.c)
+	$(call patch-filter-one,7577,gp_7500.c)
+	$(call patch-filter-one,7641,gp_7500.c)
+	$(call patch-filter-one,7642,gp_7500.c)
+	$(call patch-filter-one,7643,gp_7500.c)
+	$(call patch-filter-one,7644,gp_7500.c)
+	$(call patch-filter-one,7686,gp_7500.c)
+	$(call patch-filter-one,8025,gp_8000.c)
+	$(call patch-filter-one,8030,gp_8000.c)
+	$(call patch-filter-one,8058,gp_8000.c)
+	$(call patch-filter-one,8059,gp_8000.c)
+	$(call patch-filter-one,8066,gp_8000.c)
+	$(call patch-filter-one,8614,gp_8500.c)
+	$(call patch-filter-one,8643,gp_8500.c)
+	$(call patch-filter-one,8644,gp_8500.c)
+	$(call patch-filter-one,8673,gp_8500.c)
+	$(call patch-filter-one,8674,gp_8500.c)
+	$(call patch-filter-one,9001,gp_9000.c)
+	$(call patch-filter-one,9072,gp_9000.c)
+	$(call patch-filter-one,9073,gp_9000.c)
+	$(call patch-filter-one,9075,gp_9000.c)
+	$(call patch-filter-one,9079,gp_9000.c)
+	$(call patch-filter-one,9080,gp_9000.c)
+	$(call patch-filter-one,9082,gp_9000.c)
+	$(call patch-filter-one,9085,gp_9000.c)
+	$(call patch-filter-one,9086,gp_9000.c)
+	$(call patch-filter-one,9088,gp_9000.c)
+	$(call patch-filter-one,9090,gp_9000.c)
+	$(call patch-filter-one,9091,gp_9000.c)
+	$(call patch-filter-one,9093,gp_9000.c)
+	$(call patch-filter-one,9095,gp_9000.c)
+	$(call patch-filter-one,9096,gp_9000.c)
+	$(call patch-filter-one,9098,gp_9000.c)
+	$(call patch-filter-one,9099,gp_9000.c)
+	$(call patch-filter-one,9109,gp_9000.c)
+	$(call patch-filter-one,9110,gp_9000.c)
+	$(call patch-filter-one,9160,gp_9000.c)
+	$(call patch-filter-one,9174,gp_9000.c)
+	$(call patch-filter-one,9195,gp_9000.c)
+	$(call patch-filter-one,9196,gp_9000.c)
+	$(call patch-filter-one,9204,gp_9000.c)
+	$(call patch-filter-one,9239,gp_9000.c)
 
 .PHONY: patch
 patch: patch-openssl patch-generate-host patch-generate-ta
