+++ host/xtest/regression_7000.c
